Added a save() method on a generic Object, to the DBHelper class.
Added private property id to both Author and Book classes, plus their getters/setters.

Added a set of books that is associated with the Author class, and the corresponding getter/setter methods.

Created a simple Runner class, to instantiate a new Book and a new Author.
u
Validated DBHelper.save() worked correctly and saved the new book and author correctly to the database.